+++ php/PHPForBeginnersBecomeAPHPMasterCMSProject/cms/admin/includes/viewAllUsers.php
@@ -48,7 +48,7 @@
                 echo "<td><a href='users.php?source=edit_user&u_id=$user_id'>Edit</a></td>";
                 echo "<td><a href='users.php?change_to_admin=$user_id'>Admin</a></td>";
                 echo "<td><a href='users.php?change_to_sub=$user_id'>Subscriber</a></td>";
-                echo "<td><a href='users.php?delete=$user_id'>Delete</a></td>";
+                echo "<td><a onClick=\"javascript: return confirm('Are you sure you want to delete?');\" href='users.php?delete=$user_id'>Delete</a></td>";
                 echo "</tr>";
             }
         ?>
